UPD: reorganize conformity score tests
